feat: implement InventoryCatalog component for teacher inventory management
